requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
requests==2.31.0
beautifulsoup4==4.12.2
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.8.3
feedparser==6.0.11